Move CuratedDataLoader into python base
